src/pipelines/components/scraper_components/pd_write_embeddings/requirements.txt
fondant==0.11.2
fondant[component]
google-cloud-storage